backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
asyncpg
pydantic
python-dotenv
slowapi